+++ src/js/background/tab_state_tracker/FrameStateMachine.ts
